refactor(pharamacy): implements derived drug login in constructor
This refractoring of the constructor logic allow for now breaking the
API, user can still construct a Pharamacy with basic drugs as they used
to with the pre-refractored version of Pharamacy. But now the phamarcy
can rely on the newly created drugs internal logic to manage expiration
and benefits. Making a new drugs becomes much easier.