💥!: Drug is now abstract
Refactored the pharmacy class to remove all the logic from it.
The drugs containing their own logic, they are in charge of updating themselves.

Drug is made abstract because instantiating a drug with its name and expecting the logic to remain does not work anymore. The appropriate class should be used instead.